revisi

+++ application/controllers/admin/Data.php
@@ -1,5 +1,6 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
+date_default_timezone_set('Asia/Jakarta');
 
 class Data extends CI_Controller {
 	function __construct(){
@@ -18,6 +19,7 @@ class Data extends CI_Controller {
 		$login = $this->session->userdata('id_user');
 		$data['user'] = $this->Model_user->getLogin($login)->row();
 		$data['usr'] = $this->Model_user->getAll()->result();
+		$data['posisi'] = $this->db->query("SELECT * FROM saw_position ORDER BY id_position ASC")->result();
 		$this->load->view('layout/header',$data);
 		$this->load->view('admin/adm_set_user',$data);
 		$this->load->view('layout/footer');
@@ -38,18 +40,19 @@ class Data extends CI_Controller {
 			}else{
 				$data = array(
 					'fullname' => $this->input->post('fullname'),
+					'nip' => $this->input->post('nip'),
 					'tempat_lahir' => $this->input->post('tmpt'),
 					'tgl_lahir' => $this->input->post('tgl'),
 					'alamat' => $this->input->post('alamat'),
 					'telp' => $this->input->post('telp'),
 					'jenis_kelamin' => $this->input->post('jk'),
-					'jabatan' => $this->input->post('jb'),
+					'id_position' => $this->input->post('jb'),
 					'status' => 'on',
 					'foto' => $this->upload->data('file_name')
 				);	
 				// print_r($data);
 				$this->Model_user->simpan($data);
-				$this->session->set_flashdata('alert','ditambah');
+				$this->session->set_flashdata('success','Data berhasil ditambah');
 				redirect(base_url('admin/data'));
 			}
 		}else{
@@ -70,24 +73,36 @@ class Data extends CI_Controller {
 			if ( ! $this->upload->do_upload('foto')){
 				$data = array(
 					'fullname' => $this->input->post('fullname'),
-					'id_access' => $this->input->post('akses'),
-					'status' => $this->input->post('sts')
+					'nip' => $this->input->post('nip'),
+					'tempat_lahir' => $this->input->post('tmpt'),
+					'tgl_lahir' => $this->input->post('tgl'),
+					'alamat' => $this->input->post('alamat'),
+					'telp' => $this->input->post('telp'),
+					'jenis_kelamin' => $this->input->post('jk'),
+					'id_position' => $this->input->post('jb'),
+					'status' => $this->input->post('status'),
 				);	
 				// print_r($data);
 				$this->Model_user->update($this->input->post('ids'),$data);
-				$this->session->set_flashdata('alert','ditambah');
-				redirect(base_url('admin/setting/user'));
+				$this->session->set_flashdata('success','Data berhasil diupdate');
+				redirect(base_url('admin/data'));
 			}else{
 				$data = array(
 					'fullname' => $this->input->post('fullname'),
-					'id_access' => $this->input->post('akses'),
-					'status' => $this->input->post('sts'),
+					'nip' => $this->input->post('nip'),
+					'tempat_lahir' => $this->input->post('tmpt'),
+					'tgl_lahir' => $this->input->post('tgl'),
+					'alamat' => $this->input->post('alamat'),
+					'telp' => $this->input->post('telp'),
+					'jenis_kelamin' => $this->input->post('jk'),
+					'id_position' => $this->input->post('jb'),
+					'status' => $this->input->post('status'),
 					'foto' => $this->upload->data('file_name')
 				);	
 				// print_r($data);
 				$this->Model_user->update($this->input->post('ids'),$data);
-				$this->session->set_flashdata('alert','ditambah');
-				redirect(base_url('admin/setting/user'));
+				$this->session->set_flashdata('success','Data berhasil diupdate');
+				redirect(base_url('admin/data'));
 			}
 		}else{
 			echo "anda tidak memiliki akses halaman ini";
@@ -96,18 +111,27 @@ class Data extends CI_Controller {
 	}
 
 	public function hapus_user($id){
-		$this->Model_user->delete($id);
-		redirect(base_url('admin/setting/user'));
-	}
-
-	public function reset_password($id){
-		$reset = md5(12345);
-		$data = array(
-			'password' => $reset
-		);
-		$this->Model_user->update($id,$data);
-		redirect(base_url('admin/setting/user'));
-
+		$cek = $this->db->query("SELECT * FROM saw_evaluations WHERE id_user = '$id'")->num_rows();
+		if($cek >= 1){
+			$this->session->set_flashdata('error','Tidak dapat dihapus');
+			redirect(base_url('admin/data'));
+		}else{
+			$cek_penilai = $this->db->query("SELECT * FROM saw_login WHERE id_user = '$id'")->num_rows();
+			if($cek_penilai >= 1){
+				$this->session->set_flashdata('error','Tidak dapat dihapus');
+				redirect(base_url('admin/data'));	
+			}else{
+				$cek_alternatif = $this->db->query("SELECT * FROM saw_alternatives WHERE id_user = '$id'")->num_rows();
+				if($cek_alternatif >= 1 ){
+					$this->session->set_flashdata('error','Tidak dapat dihapus');
+					redirect(base_url('admin/data'));		
+				}else{
+					$this->Model_user->delete($id);
+					$this->session->set_flashdata('success','Data berhasil dihapus');
+					redirect(base_url('admin/data'));		
+				}
+			}
+		}
 	}
 
 }
